+++ src/Channels/PushChannel.php
@@ -3,15 +3,15 @@
 namespace Orlyapps\LaravelFirebaseNotifications\Channels;
 
 use Illuminate\Notifications\Notification;
-use Kreait\Firebase;
+use Illuminate\Support\Facades\Log;
 use Kreait\Firebase\Exception\Messaging\InvalidMessage;
 use Kreait\Firebase\Exception\Messaging\NotFound;
 use Kreait\Firebase\Exception\MessagingException;
+use Kreait\Firebase\Factory;
 use Kreait\Firebase\Messaging\AndroidConfig;
 use Kreait\Firebase\Messaging\ApnsConfig;
 use Kreait\Firebase\Messaging\CloudMessage;
 use Kreait\Firebase\Messaging\WebPushConfig;
-use Kreait\Firebase\ServiceAccount;
 
 class PushChannel
 {
@@ -27,7 +27,7 @@ class PushChannel
         $message = $notification->toPush($notifiable);
 
         foreach ($notifiable->pushTokens as $token) {
-            \Log::info('Send Push Notification to ' . $notifiable->email . ' with token' . $token);
+            Log::info('Send Push Notification to ' . $notifiable->email . ' with token' . $token);
             try {
                 $this->push($message['title'], $message['body'], $message['url'], $token->token, $message['data']);
             } catch (NotFound $e) {
@@ -35,25 +35,19 @@ class PushChannel
             } catch (InvalidMessage $e) {
                 $token->delete();
             } catch (MessagingException $e) {
-                \Log::error($e);
-                \Log::error('Failed to send Message to ' . $notifiable->email . ' with token' . $token);
+                Log::error($e);
+                Log::error('Failed to send Message to ' . $notifiable->email . ' with token' . $token);
             } catch (\Throwable $e) {
-                \Log::error($e);
+                Log::error($e);
             }
         }
-
-        // Send notification to the $notifiable instance...
     }
 
-    private function push($title, $body, $url, $deviceToken, $data)
+    protected function push($title, $body, $url, $deviceToken, $data)
     {
-        if (env('FCM_SERVICE') === null) {
-            $serviceAccount = ServiceAccount::fromValue(config('services.fcm.json_file_path'));
-        } else {
-            $serviceAccount = ServiceAccount::fromValue(env('FCM_SERVICE'));
-        }
+        $credentials = env('FCM_SERVICE') ?? config('services.fcm.json_file_path');
 
-        $firebase = (new Firebase\Factory())->withServiceAccount($serviceAccount);
+        $firebase = (new Factory())->withServiceAccount($credentials);
         $messaging = $firebase->createMessaging();
 
         $message = CloudMessage::fromArray([