+++ src/Models/PushToken.php
@@ -18,6 +18,6 @@ class PushToken extends Model
 
     public function user()
     {
-        return $this->belongsTo(\App\User::class);
+        return $this->belongsTo(config('auth.providers.users.model'));
     }
 }