+++ src/Policies/PushTokenPolicy.php
@@ -2,8 +2,8 @@
 
 namespace Orlyapps\LaravelFirebaseNotifications\Policies;
 
-use App\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
+use Illuminate\Contracts\Auth\Authenticatable;
 use Orlyapps\LaravelFirebaseNotifications\Models\PushToken;
 
 class PushTokenPolicy
@@ -11,24 +11,10 @@ class PushTokenPolicy
     use HandlesAuthorization;
 
     /**
-     * Create a new policy instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        //
-    }
-
-    /**
      * Determine whether the user can delete the model.
-     *
-     * @param  \App\User  $user
-     * @param  \App\User  $model
-     * @return mixed
      */
-    public function delete(User $user, PushToken $token)
+    public function delete(Authenticatable $user, PushToken $token)
     {
-        return $token->user_id === $user->id;
+        return $token->user_id === $user->getAuthIdentifier();
     }
 }